+++ malachite-base/src/bin_util/demo_and_bench/num/conversion/string/mod.rs
@@ -11,6 +11,7 @@ use malachite_base::test_util::runner::Runner;
 pub(crate) fn register(runner: &mut Runner) {
     from_sci_string::register(runner);
     from_string::register(runner);
+    latex::register(runner);
     options::register(runner);
     to_sci::register(runner);
     to_string::register(runner);
@@ -18,6 +19,7 @@ pub(crate) fn register(runner: &mut Runner) {
 
 mod from_sci_string;
 mod from_string;
+mod latex;
 mod options;
 mod to_sci;
 mod to_string;